requirements.txt
# Core ML Libraries
torch>=2.0.0,<2.6.0
torchvision>=0.15.0,<0.20.0
timm>=0.9.0,<1.0.0
pytorch-lightning>=2.0.0,<3.0.0
torchmetrics>=1.0.0,<2.0.0

# Data Processing
pandas>=1.5.0,<3.0.0
numpy>=1.24.0,<2.0.0
opencv-python>=4.7.0,<5.0.0
Pillow>=9.5.0,<11.0.0
albumentations>=1.3.0,<2.0.0

# Visualization & Monitoring
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0
plotly>=5.14.0,<6.0.0
wandb>=0.15.0,<1.0.0
tensorboard>=2.13.0,<3.0.0

# Scientific Computing
scipy>=1.10.0,<2.0.0
scikit-learn>=1.2.0,<2.0.0
scikit-image>=0.20.0,<1.0.0

# Medical Image Processing
pydicom>=2.3.0,<3.0.0
SimpleITK>=2.2.0,<3.0.0

# Configuration & Utilities
tqdm>=4.65.0,<5.0.0
pyyaml>=6.0,<7.0.0
omegaconf>=2.3.0,<3.0.0
hydra-core>=1.3.0,<2.0.0
rich>=13.0.0,<14.0.0

# Development & Jupyter
jupyter>=1.0.0,<2.0.0
ipywidgets>=8.0.0,<9.0.0
notebook>=6.5.0,<8.0.0
jupyterlab>=4.0.0,<5.0.0

# Model Interpretation & Explainability
grad-cam>=1.4.0,<2.0.0
captum>=0.6.0,<1.0.0

# Additional Utilities
fastapi>=0.100.0,<1.0.0  # For model serving API
uvicorn>=0.20.0,<1.0.0   # ASGI server for FastAPI
python-multipart>=0.0.6,<1.0.0  # File uploads for API
requests>=2.30.0,<3.0.0  # HTTP requests
